Changed Face() and made the actual generation of a Rubik's Cube
- Setting squares of Face() is now simplified

- scramble() exists, which, when given 6 faces, generates 9 instances of each colour (currently represented as a number) for each of the 54 squares visible

NOTE : yes I know rubik's is misspelt as 'rubix'

NOTE : The current code generating a random rubik's cube will, most of the time, generate a cube that is impossible to solve. This is something that i've decided to just deal with, and maybe when i'm a better programmer I can come back to this and add a better method ( maybe generating a solved cube and randomly turning it or something ), but I'm not that programmer yet.